refactor: fix notifications legacy FK + move TelegramSubscription
Data model audit WS-E:

- Remove truth_social_shitposts JOIN from get_new_predictions_since(),
  use predictions.post_timestamp instead (fixes silent miss of future
  signal-based predictions)
- Move TelegramSubscription from shitvault/shitpost_models.py to
  notifications/models.py (proper separation of concerns)
- Re-export from shitpost_models.py for backward compatibility
